add {'0'} NULL at the end
Converter function works well that is written in C but it does print some special or ambiguous characters at the end. There after long search and hassle this is the fix that did the trick, which tell converter to stop at Null or not to include unnecessary characters, as we are anyways assuming all strings with 512 size.